DeleteDepartamento

+++ src/components/DeleteDepartamento.js
@@ -2,33 +2,76 @@ import React, { Component } from "react";
 import axios from "axios";
 import Global from "./Global";
 import loading from "./../assets/images/loading.jpg";
-import { NavLink } from "react-router-dom";
+import { Navigate, NavLink } from "react-router-dom";
+import Swal from "sweetalert2"; // Asegúrate de importar SweetAlert
 
 export default class DeleteDepartamento extends Component {
 
   state = {
     status: false,
+    completado: false,
+    departamento: null
   };
 
-  loadDepartamento = () => {
-    
-  }
-
-  deleteDepartamento = () => {
-    let id = this.props.id
-    let request = "api/departamentos/" + id;
+  findDepartamento = () => {
+    let request = "api/departamentos/" + this.props.id; //el id viene dentro de props
     let url = Global.apiUrlDepartamentos + request;
 
-    axios.delete(url).then((response) => {
-      console.log("Eliminado");
+    axios.get(url).then((response) => {
+      console.log("Detalles departamento");
+      this.setState({
+        departamento: response.data,
+        status: true
+      });
+    });
+  };
+
+  componentDidMount = () => {
+    this.findDepartamento();
+  };
+
+  deleteDepartamento = () => {
+    // Mostrar alerta de confirmación antes de eliminar
+    Swal.fire({
+      title: "¿Estás seguro?",
+      text: "¡No podrás revertir esto!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Sí, ¡elimínalo!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        let id = this.props.id;
+        let request = "api/departamentos/" + id;
+        let url = Global.apiUrlDepartamentos + request;
+
+        // Si se confirma, realiza la eliminación
+        axios.delete(url).then((response) => {
+          console.log("Eliminado " + id);
+          this.setState({
+            completado: true
+          });
+
+          // Mostrar alerta de éxito
+          Swal.fire({
+            title: "¡Eliminado!",
+            text: "El departamento ha sido eliminado.",
+            icon: "success"
+          });
+        });
+      }
     });
   };
 
   render() {
+    if (this.state.completado === true) {
+      return (<Navigate to="/" />);
+    } 
     return (
       <div>
         <NavLink to="/">Back to List</NavLink>
-        {this.state.status == true ? (
+        {this.state.status === true ? (
           <div>
             <ul className="list-group-item">
               <li className="list-group-item">
@@ -57,7 +100,7 @@ export default class DeleteDepartamento extends Component {
               transform: "translate(-50%, -50%)",
             }}
           />
-        )}
+        )}   
       </div>
     );
   }
